serverless_chat/requirements.txt
aws-cdk-lib==2.206.0
constructs>=10.0.0,<11.0.0
boto3>=1.20.0,<2.0.0
botocore>=1.20.0,<2.0.0
PyJWT>=2.0.0,<3.0.0
requests>=2.25.0,<3.0.0
